// server/src/enrichUser/interface.ts
import { Gender } from "./index";
import { Optional } from "sequelize";

interface IUser {
  gender: string;
  name: {
    title: string;
    first: string;
    last: string;
  };
  location: {
    street: {
      number: number;
      name: string;
    };
    city: string;
    state: string;
    country: string;
    postcode: string;
    coordinates: {
      latitude: string;
      longitude: string;
    };
    timezone: {
      offset: string;
      description: string;
    };
  };
  email: string;
  login: {
    uuid: string;
    username: string;
    password: string;
    salt: string;
    md5: string;
    sha1: string;
    sha256: string;
  };
  dob: {
    date: string;
    age: number;
  };
  registered: {
    date: string;
    age: number;
  };
  phone: string;
  cell: string;
  id: {
    name: string;
    value: string;
  };
  picture: {
    large: string;
    medium: string;
    thumbnail: string;
  };
  nat: string;
}

interface IRandomUserResponse {
  results: IUser[];
  info: {
    seed: string;
    results: number;
    page: number;
    version: string;
  };
}

interface IGuessGenderResponse {
  name: string;
  gender: Gender;
  probability: number;
  count: number;
}

interface IuserAttributes {
  id: number;
  name: string;
  gender?: string;
  address?: string;
  email?: string;
  dob?: string;
  age?: number;
  phone?: string;
  score?: number;
}
interface IuserInput extends Optional<IuserAttributes, "id"> {}

export {
  IUser,
  IRandomUserResponse,
  IGuessGenderResponse,
  IuserInput,
  IuserAttributes,
};